Remove unneeded CMake settings (assume at least open62541 v1.3)

+++ backends/stdout/examples/dump.c
@@ -20,7 +20,7 @@ printId(const UA_NodeId *id) {
     return nodeidDump;
 }
 
-int addNamespace(void *userContext, const char *uri) { return 1; }
+unsigned short addNamespace(void *userContext, const char *uri) { return 1; }
 
 void dumpNode(void *userContext, const NL_Node *node)
 {
